Introduce a snapshots#show page
We needed to add a way to identify individual snapshots. We added a snapshot slug,
which consists of two parts - a test case file path, and an index. The index maps to the numbers of times the #take_snapshot method was called in a single test case.

An example of a slug: `minitest/snapshot/snapshots_controller_test/test_should_get_index/0`

With a use of a slug we can link from a snapshots#index page to an individual snapshots#show page where we render a body of the captured response.