Order snapshots by a source location
That way the order in which the snapshots are displayed maps to the order in which
they are called in test files.

As a secondary order we use a slug. That way if multiple snapshots are taken in the
same test case, they are guaranteed to be ordered by how they were taken.