Autoload routes after initialize
It reduces the number of steps we needs to make to start using the engine.

We use /rails/* prefix in the URL path since it ensures there will be no clashes with other 3rd party libraries.

It only mounts the routes in development environment. We don't need them in test environment since a webserver isn't running in that time.