WIP (needs soft reset before continuing): show snapshot's request and response details
- controller and action of a request
- response status
- all request headers
- all response headers

Exploring if it is useful to collect and display more details.

A future idea: I could have two representations of data. The current one which is test case oriented, and another one which is controller action oriented. For each controller action there would listed examples of requests/responses based on the taken snapshots.