Speed up a browser refresh when live reloading
The change removes flickering of a browser page while tests are running.

### How things worked before the introduced change

When live reloading was enabled, and a user opened /rails/snapshots page in a browser, and ran the tests for taking snapshots the following happened

1) tmp snapshot directory was cleaned
2) tests started to run, and snapshots were being taken
3) in the mean time a browser reloaded the /rails/snapshots page showing an empty list of snapshots
4) when tests were complete (or even before) reloads of the /rails/snapshots page happened showing newly generated snapshots

### How things work with the introduced change

Instead of 1 we have 2 tmp directories. /tmp/view_processing/snapshots and /tmp/view_processing/processing

When live reloading is enabled, and a user opens /rails/snapshots page in a browser, and runs the tests for taking snapshots the following happens

1) the processing directory is cleaned
2) tests start to run, and snapshots are being taken and written to the processing directory
4) when tests are complete the storage directory is cleaned and we move the snapshots from the processing directory to the snapshots directory.
5) the browser reloads the /rails/snapshots page showing an updated list of snapshots