Remove traces of javascript in a snapshot's html when rendering
That way we see the page the way it looks as if the javascript would be disabled in a browser.

It also prevents any manipulation of the DOM that could mislead us when we inspect it.

The page will be seen as integration tests will see it (minus the javascript bits).

We should allow to toggle this feature by a user.