A bigger refactor of the domain model
1) Rename TestRecording into Context

We no longer talk about a snapshot's test recording. Instead, we talk about a context in which a snapshot is made. Later on the context will probably have a source that will be either minitest or rspec.

2) Rename MailRecording and ResponseRecording into a MailType and ResponseType

The idea is to deal only with one class Snapshot that is extended with a type specific behaviour. Snapshot's type is determined by a snapshotee class. For each snapshotee class a library user needs to introduce a type that knows how to handle the snapshotee. To that end we introduce an abstract Snapshot::Type that individual implementations of types need to inherit from. a type implementation at minimum needs to implement #extract and #from_hash methods. #extract defines what attributes we are interested in snapshotting from a given snapshotee. #from_hash deals with deserializing snapshot's data from storage. Instance methods from a type implementation are delegated to a Snapshot instance.

3) Simplify structure of a persisted snapshot's context

3 attributes are no longer persisted since they can be constructed at read time.

4) Fix a bug in Snapshots::MailRecordingsController

- It wasn't possible to open a mail snapshot in a text format.
- There wasn't a default mime type when rendering a raw mail.