Rename the library from ViewInspector to SnapshotInspector
The name View Inspector didn't feel quite right, since we are not inspecting views directly, but the snapshots of the outputs of the views - responses, mail messages, and similar.

Snapshot Inspector feels as a very straightforward name. We are taking snapshots and inspecting them in a browser. Since it uses a suffix *Inspector the monotone repetition of a word "Snapshot" in class names is eased. For example, SnapshotInspector::Snapshot sounds pretty ok.