Move responsibility over detecting snapshot type to SnapshotInspector::Snapshot::Type
Snapshot shouldn't be responsible for figuring out mapping between snapshotee classes and types. We define the mapping in the context of type implementations, so everything related to that should be contained within the Type class.

We needed to restructure the persisted JSON structure a bit.

We override Type#as_json in order to avoid defining an instance variable @snapshotee_class on a level of a specific type implementation.

That way type implementations don't have to be aware of what we do with snapshotee, except for the part when define it.